refactor(api): extract admin activity logic to service
- Move business logic from route handler to adminMetricsService
- Add ConsentDocument type to types/firestore.ts
- Remove duplicate ConsentDocument from userService
