+++ src/app/api/admin/activity/route.ts
@@ -5,110 +5,43 @@
  */
 import { type NextRequest, NextResponse } from "next/server";
 import { verifyAdminTokenWithPermission } from "@/lib/adminAuth";
-import { db } from "@/lib/firebaseAdmin";
-import { getTodayStartColombia } from "@/lib/utils/dateUtils";
+import { createLogger } from "@/lib/logger";
+import { adminMetricsService } from "@/services/adminMetricsService";
+
+const logger = createLogger("ApiAdminActivity");
 
 export async function GET(request: NextRequest) {
 	try {
-		// Verificar autenticación y permiso dashboard:view
-		const authResult = await verifyAdminTokenWithPermission(request, "dashboard:view");
+		const authResult = await verifyAdminTokenWithPermission(
+			request,
+			"dashboard:view",
+		);
 		if (!authResult.success) {
 			return authResult.response;
 		}
 
-		// Fecha de inicio del día (medianoche hora Colombia - compatible con UTC del servidor)
 		const now = new Date();
-		const todayStart = getTodayStartColombia();
-
-		// OPTIMIZADO: Ejecutar queries en paralelo y usar .select() para solo traer campos necesarios
-		// Agregado limit(1000) como seguridad para evitar lecturas masivas
-		const [todaySnapshot, latestSnapshot] = await Promise.all([
-			// Consentimientos de hoy - solo campos necesarios para stats
-			db
-				.collection("consents")
-				.where("signedAt", ">=", todayStart)
-				.select("signedAt", "minorsSnapshot")
-				.limit(1000) // Límite de seguridad
-				.get(),
-			// Últimos 10 consentimientos - solo campos necesarios para el feed
-			db
-				.collection("consents")
-				.orderBy("signedAt", "desc")
-				.limit(10)
-				.select("consecutivo", "signedAt", "adultSnapshot", "minorsSnapshot")
-				.get(),
-		]);
-
-		const consentsToday = todaySnapshot.size;
-
-		// Calcular total de menores hoy
-		let minorsToday = 0;
-		const hourlyStats: Record<number, number> = {};
-
-		todaySnapshot.docs.forEach((doc) => {
-			const data = doc.data();
-			minorsToday += data.minorsSnapshot?.length || 0;
-
-			// Estadísticas por hora
-			if (data.signedAt) {
-				const signedDate =
-					data.signedAt instanceof Date
-						? data.signedAt
-						: data.signedAt.toDate();
-				const hour = signedDate.getHours();
-				hourlyStats[hour] = (hourlyStats[hour] || 0) + 1;
-			}
-		});
-
-		const latestConsents = latestSnapshot.docs.map((doc) => {
-			const data = doc.data();
-
-			// Convertir timestamp a ISO string
-			let signedAt = null;
-			if (data.signedAt) {
-				if (data.signedAt instanceof Date) {
-					signedAt = data.signedAt.toISOString();
-				} else if (typeof data.signedAt.toDate === "function") {
-					signedAt = data.signedAt.toDate().toISOString();
-				}
-			}
-
-			return {
-				id: doc.id,
-				consecutivo: data.consecutivo,
-				adultName: data.adultSnapshot?.fullName || "N/A",
-				minorsCount: data.minorsSnapshot?.length || 0,
-				signedAt,
-			};
-		});
-
-		// Convertir a array ordenado (solo horario comercial 8am-10pm)
-		const hourlyData = Array.from({ length: 24 }, (_, hour) => ({
-			hour,
-			label: `${hour.toString().padStart(2, "0")}:00`,
-			count: hourlyStats[hour] || 0,
-		})).filter((h) => h.hour >= 8 && h.hour <= 22);
+		const activity = await adminMetricsService.getAdminActivity();
 
 		return NextResponse.json(
 			{
 				success: true,
 				stats: {
-					consentsToday,
-					minorsToday,
+					consentsToday: activity.consentsToday,
+					minorsToday: activity.minorsToday,
 					timestamp: now.toISOString(),
 				},
-				latestConsents,
-				hourlyData,
+				latestConsents: activity.latestConsents,
+				hourlyData: activity.hourlyData,
 			},
 			{
 				headers: {
-					// Cache por 30 segundos para reducir lecturas de Firestore
 					"Cache-Control": "private, max-age=30, stale-while-revalidate=60",
 				},
-			}
+			},
 		);
 	} catch (error) {
-		console.error("[API /admin/activity] Error:", error);
+		logger.error("Error obteniendo actividad", error);
 		return NextResponse.json(
 			{ error: "Error al obtener actividad" },
 			{ status: 500 },
